+++ test/integration/serverTest.js
@@ -15,6 +15,7 @@ const db = imports.db
 const redis = imports.redis
 const server = imports.server
 const braveAlerter = imports.braveAlerter
+const particle = imports.particle
 const StateMachine = require('../../stateMachine/StateMachine')
 const XETHRU_STATE = require('../../SessionStateXethruEnum')
 const SENSOR_EVENT = require('../../SensorEventEnum')
@@ -38,8 +39,9 @@ const door_coreID = 'door_particlecoreid1'
 const radar_coreID = 'radar_particlecoreid1'
 
 async function firmwareAlert(coreID, sensorEvent) {
+  let response
   try {
-    await chai.request(server).post('/api/sensorEvent').send({
+    response = await chai.request(server).post('/api/sensorEvent').send({
       event: sensorEvent,
       data: 'test-event',
       ttl: 60,
@@ -50,6 +52,7 @@ async function firmwareAlert(coreID, sensorEvent) {
   } catch (e) {
     helpers.log(e)
   }
+  return response
 }
 
 async function innosentMovement(coreID, min, max) {
@@ -154,6 +157,148 @@ describe('Brave Sensor server', () => {
     await redis.disconnect()
   })
 
+  describe('POST request: alerts from firmware state machine for locations with a non-null Particle Siren Id', () => {
+    beforeEach(async () => {
+      await redis.clearKeys()
+      await db.clearSessions()
+      await db.clearLocations()
+      await db.clearClients()
+      const client = await clientFactory(db)
+      await db.createLocation(
+        testLocation1Id,
+        testLocation1PhoneNumber,
+        MOVEMENT_THRESHOLD,
+        STILLNESS_TIMER,
+        DURATION_TIMER,
+        1000,
+        INITIAL_TIMER,
+        ['+15005550006'],
+        '+15005550006',
+        ['+15005550006'],
+        1000,
+        'locationName',
+        door_coreID,
+        radar_coreID,
+        'XeThru',
+        'alertApiKey',
+        true,
+        true,
+        'particleCoreIdTest',
+        client.id,
+      )
+      sandbox.stub(braveAlerter, 'startAlertSession')
+      sandbox.stub(braveAlerter, 'sendSingleAlert')
+      sandbox.spy(particle, 'callFunction')
+      sandbox.spy(helpers, 'logError')
+    })
+
+    afterEach(async () => {
+      await redis.clearKeys()
+      await db.clearSessions()
+      await db.clearLocations()
+      await db.clearClients()
+      sandbox.restore()
+      helpers.log('\n')
+    })
+
+    it('should return 200 to a request with no headers', async () => {
+      const response = await chai.request(server).post('/api/sensorEvent').send({})
+      expect(response).to.have.status(200)
+    })
+
+    it('should log an error for a request with no headers', async () => {
+      await chai.request(server).post('/api/sensorEvent').send({})
+      expect(helpers.logError).to.have.been.called
+    })
+
+    it('should return 200 for a valid request', async () => {
+      const response = await firmwareAlert(radar_coreID, SENSOR_EVENT.STILLNESS)
+      expect(response).to.have.status(200)
+    })
+
+    it('should call particle.callFunction with the correct device ID if sirenParticleId is not null', async () => {
+      await firmwareAlert(radar_coreID, SENSOR_EVENT.STILLNESS)
+      expect(particle.callFunction).to.have.been.calledWith({
+        deviceId: 'particleCoreIdTest',
+        name: 'start',
+        argument: 'start',
+        auth: helpers.getEnvVar('PARTICLE_ACCESS_TOKEN'),
+      })
+    })
+  })
+
+  describe('POST request: alerts from firmware state machine for locations with a null Particle Siren Id', () => {
+    beforeEach(async () => {
+      await redis.clearKeys()
+      await db.clearSessions()
+      await db.clearLocations()
+      await db.clearClients()
+      const client = await clientFactory(db)
+      await db.createLocation(
+        testLocation1Id,
+        testLocation1PhoneNumber,
+        MOVEMENT_THRESHOLD,
+        STILLNESS_TIMER,
+        DURATION_TIMER,
+        1000,
+        INITIAL_TIMER,
+        ['+15005550006'],
+        '+15005550006',
+        ['+15005550006'],
+        1000,
+        'locationName',
+        door_coreID,
+        radar_coreID,
+        'XeThru',
+        'alertApiKey',
+        true,
+        true,
+        null,
+        client.id,
+      )
+      sandbox.stub(braveAlerter, 'startAlertSession')
+      sandbox.stub(braveAlerter, 'sendSingleAlert')
+      sandbox.spy(particle, 'callFunction')
+      sandbox.spy(helpers, 'logError')
+    })
+
+    afterEach(async () => {
+      await redis.clearKeys()
+      await db.clearSessions()
+      await db.clearLocations()
+      await db.clearClients()
+      sandbox.restore()
+      helpers.log('\n')
+    })
+
+    it('should return 200 to a request with no headers', async () => {
+      const response = await chai.request(server).post('/api/sensorEvent').send({})
+      expect(response).to.have.status(200)
+    })
+
+    it('should log an error for a request with no headers', async () => {
+      await chai.request(server).post('/api/sensorEvent').send({})
+      expect(helpers.logError).to.have.been.called
+    })
+
+    it('should return 200 for a valid request', async () => {
+      const response = await firmwareAlert(radar_coreID, SENSOR_EVENT.STILLNESS)
+      expect(response).to.have.status(200)
+    })
+
+    it('should not call particle.callFunction if there sirenParticleId is null', async () => {
+      await firmwareAlert(radar_coreID, SENSOR_EVENT.STILLNESS)
+      expect(particle.callFunction).to.have.not.been.called
+    })
+
+    it('should create a session with the right alert type if particleSirenId is null', async () => {
+      await firmwareAlert(radar_coreID, SENSOR_EVENT.STILLNESS)
+      const sessions = await db.getAllSessionsFromLocation(testLocation1Id)
+      const session = sessions[sessions.length - 1]
+      expect(session.alertType).to.equal(ALERT_TYPE.SENSOR_STILLNESS)
+    })
+  })
+
   describe('POST request: alerts from firmware state machine', () => {
     beforeEach(async () => {
       await redis.clearKeys()
@@ -181,6 +326,7 @@ describe('Brave Sensor server', () => {
         'alertApiKey',
         true,
         true,
+        null,
         client.id,
       )
       sandbox.stub(braveAlerter, 'startAlertSession')
@@ -278,6 +424,7 @@ describe('Brave Sensor server', () => {
         'alertApiKey',
         true,
         false,
+        null,
         client.id,
       )
       sandbox.stub(braveAlerter, 'startAlertSession')
@@ -400,6 +547,7 @@ describe('Brave Sensor server', () => {
         'alertApiKey',
         false,
         false,
+        null,
         client.id,
       )
       await im21Door(door_coreID, im21door.createClosedSignal())
@@ -481,6 +629,7 @@ describe('Brave Sensor server', () => {
         'alertApiKey',
         true,
         false,
+        null,
         client.id,
       )
       await im21Door(door_coreID, im21door.createClosedSignal())
@@ -604,6 +753,7 @@ describe('Brave Sensor server', () => {
         'alertApiKey',
         false,
         false,
+        null,
         client.id,
       )
       await im21Door(door_coreID, im21door.createClosedSignal())
@@ -682,6 +832,7 @@ describe('Brave Sensor server', () => {
           'alertApiKey',
           true,
           false,
+          null,
           client.id,
         )
         await im21Door(door_coreID, im21door.createClosedSignal())
@@ -751,6 +902,7 @@ describe('Brave Sensor server', () => {
           'alertApiKey',
           true,
           false,
+          null,
           client.id,
         )
         await im21Door(door_coreID, im21door.createClosedSignal())